added support to User get API(Good to have) created separate file for Schemas and Models to resolved cyclic dependency
