--- models/__init__.py
@@ -1,4 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-
-db = SQLAlchemy()
