+++ models_schemas/__init__.py
@@ -1,0 +1,7 @@
+from flask_marshmallow import Marshmallow
+from flask_sqlalchemy import SQLAlchemy
+from flask_marshmallow import Marshmallow
+
+
+db = SQLAlchemy()
+ma = Marshmallow()
