+++ models_schemas/models/user_model.py
@@ -1,11 +1,11 @@
-from flask_sqlalchemy import SQLAlchemy
-from flask_marshmallow import Marshmallow
-from marshmallow import Schema, fields, validate, ValidationError, post_load
+# from flask_sqlalchemy import SQLAlchemy
+# from flask_marshmallow import Marshmallow
 from datetime import datetime
 
 # from models.restaurant import Restaurant
 
-from models import db
+from models_schemas import db
+# from models_schemas.schemas.restaurant_schema import RestaurantSchema
 # db = SQLAlchemy()
 
 class User(db.Model):
@@ -52,32 +52,3 @@ class User(db.Model):
     def get_id(self):
         return self.__id
 
-class UserSchema(Schema):
-    # __id = fields.Int()
-    id = fields.Int(load_only=True)
-    name = fields.Str(validate=validate.Length(min=2))
-    email = fields.Str(validate=validate.Email())
-    type = fields.Str(validate=validate.OneOf(["normal-user", "restraunt-owner"]))
-    created_at = fields.Str()
-    city = fields.Str(validate=validate.Length(min=2))
-    Zipcode = fields.Int()
-    Balance = fields.Int(validate=validate.Range(min=0))
-    status = fields.Bool(load_only=True)
-    
-    # class Meta:
-    #     fields = ("name", "email")
-    #     model = User
-
-    @post_load
-    def make_user(self,data, **kwargs):
-        return User(**data)
-
-
-# in_date = {"name":"a", "permission":"admin", "age":21}
-# try:
-#     print(UserSchema().load(in_date))
-# except ValidationError as err:
-#     print(err.messages)
-
-user_schema = UserSchema()
-users_schema = UserSchema(many=True)
